requirements.txt
absl-py==1.0.0
aniso8601==9.0.1
asgiref==3.5.0
astunparse==1.6.3
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.1
colorama==0.4.4
cycler==0.11.0
dj-database-url==0.5.0
Django==3.2.3
django-cors-headers==3.10.0
django-heroku==0.3.1
djangorestframework==3.12.4
flatbuffers==2.0
fonttools==4.29.0
gast==0.4.0
google-auth==2.4.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.43.0
gunicorn==20.1.0
h5py==3.6.0
idna==3.3
imageio==2.14.1
importlib-metadata==4.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
libclang==12.0.0
lxml==4.7.1
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
mysqlclient==2.0.3
networkx==2.6.3
numpy==1.22.1
oauthlib==3.1.1
opencv-python-headless
opt-einsum==3.3.0
packaging==21.3
Pillow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.2.0
regex==2021.4.4
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
scipy==1.7.3
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.29
sqlparse==0.4.2
tensorflow-cpu==2.7.0
termcolor==1.1.0
threadpoolctl==3.0.0
tifffile==2021.11.2
tqdm==4.61.0
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.8
Werkzeug==2.0.2
whitenoise==5.3.0
wrapt==1.13.3
zipp==3.7.0
